token based login system is added. also valid token is required to create, update & delete any record

+++ routes/income/incomeRecordsRouter.js
@@ -1,52 +1,68 @@
 const incomeRecordsRouter = require('express').Router()
 const incomeRecordsSchema = require('../../schema/income/incomeRecordsSchema')
+const { getTokenFromRequest } = require('../../helper/getTokenFromRequest')
+const jwt = require('jsonwebtoken')
+const SECRETKEY = process.env.SECRETKEY
+const userSchema = require('../../schema/user/userSchema')
 
-incomeRecordsRouter.get('/', async (req, res) => {
+incomeRecordsRouter.get('/:userId', async (req, res) => {
+  const userId = req.params.userId
+
   try {
-    const allIncomeRecords = await incomeRecordsSchema.find({})
-    console.log('Fetched all income records!')
-    res.json(allIncomeRecords)
+    const user = await userSchema.findById(userId)
+    console.log('user exists in database')
+    
   } catch (e) {
-    console.log(e)
-    res.status(400).json({
-      error: 'error fetching all income records'
+    console.log('user not found in db')
+    res.status(400).end({
+      error: 'user not found in db'
     })
+    return
   }
-})
-
-incomeRecordsRouter.get('/:id', async (req, res) => {
-  const id = req.params.id
 
   try {
-    const fetchedRecord = await incomeRecordsSchema.findById(id)
-    if (fetchedRecord) {
-      console.log('Requested record fetched!')
-      res.json(fetchedRecord)
-    } else {
-      console.log('Requested record not found!!!')
-      res.json({
-        error: 'requested record not found'
-      })
-    }
+    const fetchedRecord = await incomeRecordsSchema.find({userId})
+
+    console.log('Records fetched for user')
+    res.json(fetchedRecord)
+
   } catch (e) {
-    console.log(e)
+    console.log('error fetching records')
     res.status(400).json({
-      error: 'error fetching requested record'
+      error: 'error fetching records'
     })
   }
 })
 
 incomeRecordsRouter.post('/', async (req, res) => {
-  const sentRecord = req.body
-  const newRecord = new incomeRecordsSchema(sentRecord)
+  let decodedToken = ''
+  try {
+    const token = await getTokenFromRequest(req)
+    decodedToken = await jwt.verify(token, SECRETKEY)
+    
+
+  } catch (e) {
+    console.log('token expired or not valid')
+    res.status(400).json({
+      error: 'token expired or not valid'
+    })
+    return
+  }
+
+  const newRecord = new incomeRecordsSchema({
+    ...req.body, 
+    userId: decodedToken.userId
+  })
+
   try {
     const savedRecord = await newRecord.save()
-    console.log('New income record saved!')
+    console.log(`New income record saved by user '${decodedToken.username}'`)
     res.status(201).json(savedRecord)
+
   } catch (e) {
-    console.log(e)
+    console.log('error saving new record to db')
     res.status(400).json({
-      error: 'error saving new income record'
+      error: 'error saving new record to db'
     })
   }
 })
@@ -55,15 +71,56 @@ incomeRecordsRouter.delete('/:id', async (req, res) => {
   const id = req.params.id
   
   try {
-    await incomeRecordsSchema.findByIdAndDelete(id)
-    console.log('Record deleted!')
-    res.status(204).end()
+    const recordToDelete = await incomeRecordsSchema.findById(id)
+
+    if (recordToDelete) {
+      console.log('record exists in database')
+    }
   } catch (e) {
-    console.log(e)
+    console.log('record does not exist in database')
     res.status(400).json({
-      error: 'error deleting the record'
+      error: 'record does not exist in database'
+    })
+    return
+  }
+
+  let decodedToken = ''
+  try {
+    const token = await getTokenFromRequest(req)
+    decodedToken = await jwt.verify(token, SECRETKEY)
+
+  } catch (e) {
+    console.log('token expired or not valid')
+    res.status(400).json({
+      error: 'token expired or not valid'
+    })
+    return
+  }
+
+  const recordToDelete = await incomeRecordsSchema.findById(id)
+
+  if (recordToDelete.userId === decodedToken.userId) {
+    
+    try {
+      await incomeRecordsSchema.findByIdAndDelete(id)
+      console.log('Record deleted!')
+      res.status(204).end()
+
+    } catch (e) {
+      console.log('error deleting the record from database')
+      res.status(400).json({
+        error: 'error deleting the record from database'
+      })
+      return
+    }
+
+  } else {
+    console.log('Not authorized to delete!')
+    res.status(400).json({
+      error: 'Not authorized to delete!'
     })
   }
+
 })
 
 incomeRecordsRouter.patch('/:id', async (req, res) => {
@@ -75,8 +132,9 @@ incomeRecordsRouter.patch('/:id', async (req, res) => {
     console.log('Requested record updated!')
     const updatedRecord = await incomeRecordsSchema.findById(id)
     res.json(updatedRecord)
+
   } catch (e) {
-    console.log(e)
+    console.log('error updating record')
     res.status(400).json({
       error: 'error updating record'
     })
